Fix signature of JsonValue#object
In order to avoid writing some casts, the signature of `JsonValue#object` has to be updated.

So we can now write without cast :
```java
Map<String, Object> object = object();
// fill in the object
// ...
return json(object);
```
That avoids to know that the returned instance of `Map` is actually a `LinkedHashMap`.
The other methods (`array` and `set`) have a correct return type.