CREST-402 Default values in JsonSchema can be any JSON type
Trying to use a non-string "default" property in a JSON schema makes OpenApiTransformer.buildProperty() throw a JsonValueException.

When building an Open API property, the code assumes that a default value is a string,
 which is enforced by the call to JsonValue.asString(),
 but it can be any valid json type, including boolean, integer, array, etc.

OpenApiTransformer.java:
In buildProperty(), changed the code to convert the value to a String without using JsonValue.asString().