Fix HMAC Race Conditions in `SecureCsvWriterTest` (#6)
The issue here appears to be that if tests would take longer than 5 seconds to run (e.g. on a slower machine) they end up with HMAC signatures before the tests get a chance to check the actual audit output against the expected output.

Based on a hint provided in af741daae94cc7138e2d061fb71029b81524d79d by maximthomas from the OpenIdentityPlatform, it appears that the root cause of this issue is actually the rotation retention thread. The default rotation interval is 5 seconds, and a signature is forcibly written to a file that is about to be rotated. For this reason, it would seem that we need to extend out the rotation retention interval when extending out the signature interval. This does, in fact, appear to resolve the issue.

Closes #6.