Handle Password PIN Files with POSIX Newlines on Windows
If the repo is checked-out with `core.autocrlf` set to `input` or `false`, then the platform-specific newline doesn't match what the PIN file contains. Similarly, if a user is editing a password PIN file on Windows using an editor configured to add POSIX-compliant, trailing line endings, they would also have encountered this issue on Windows.