Remove Unnecessary Casts/Drop Eclipse Compiler from FR Util
It turns out that these casts truly are unnecessary, but that the Eclipse JDK 8 compiler has a bug that causes it not to infer upper bounds correctly.

Likely this issue:
https://bugs.eclipse.org/bugs/show_bug.cgi?id=497905#c1

Instead of the pushing forward with the unnecessary casts, I'm dropping the Eclipse compiler in favor of javac which infers the types properly.