Ensure Commons Tests Don't Fail when Surefire is Single-threaded
When debugging tests, it's common to use `-DforkCount=0`. But, if Surefire never forks, tests inherit the classpath and working directory of the parent Maven process instead of starting from a classpath and working directory that's specific to the POM module being tested. This caused several tests to fail when being debugged that otherwise pass when Surefire is run with parallel tests enabled.

Now, we simply use information Surefire passes us on the command line to ensure that resources are always loaded from a path relative to the module directory instead of the project root.

For `ScriptEngineManager`, it would appear that we need to pass-in `null` in order for the system-default engines to be made available regardless of which classpath the test has inherited. See: https://stackoverflow.com/a/25332851/4342230